refactor(offline): promote OfflineIndicator + OfflineGated to directory modules
Per module checklist: a file with a companion `.test.tsx` sibling, or
a tightly-coupled helper (`.module.css`), should live in a directory
module named after the component.

- OfflineIndicator.{tsx,test.tsx} -> OfflineIndicator/OfflineIndicator.{tsx,test.tsx}
- OfflineGated.{tsx,module.css}   -> OfflineGated/OfflineGated.{tsx,module.css}
- OfflineUnavailableTooltipLabel stays flat (no companion files).

Also drops the "Notion-style" phrasing from OfflineUnavailableTooltipLabel's
docstring per review.
