+++ src/components/offline/OfflineIndicator/OfflineIndicator.test.tsx
@@ -1,7 +1,7 @@
 import { I18nProvider } from "@lingui/react";
 import { MantineProvider } from "@mantine/core";
 import { afterEach, describe, expect, it, vi } from "vitest";
-import { OfflineIndicator } from "@/components/offline/OfflineIndicator";
+import { OfflineIndicator } from "@/components/offline/OfflineIndicator/OfflineIndicator";
 import { i18n } from "@/i18n/i18n";
 import { render, screen } from "@/test-utils";
