+++ src/components/offline/OfflineUnavailableTooltipLabel.tsx
@@ -2,7 +2,7 @@ import { Trans } from "@lingui/react/macro";
 import { Stack, Text } from "@mantine/core";
 
 /**
- * Standard Notion-style offline tooltip copy for gated controls and links.
+ * Standard offline tooltip copy for gated controls and links.
  */
 export function OfflineUnavailableTooltipLabel(): JSX.Element {
   return (
